+++ route_go/route/view_random.go
@@ -1,0 +1,32 @@
+package route
+
+import (
+	"database/sql"
+	"encoding/json"
+	"fmt"
+
+	"opennamu/route/tool"
+)
+
+func View_random(call_arg []string) {
+	db_set := map[string]string{}
+	json.Unmarshal([]byte(call_arg[0]), &db_set)
+
+	db := tool.DB_connect(db_set)
+	if db == nil {
+		return
+	}
+	defer db.Close()
+
+	var title string
+	err := db.QueryRow(tool.DB_change(db_set, "select title from data where title not like 'user:%' and title not like 'category:%' and title not like 'file:%' order by random() limit 1")).Scan(&title)
+	if err != nil {
+		if err == sql.ErrNoRows {
+			title = ""
+		} else {
+			return
+		}
+	}
+
+	fmt.Print(title)
+}
